refactor: organize app and config module
