+++ src/config/AppConfigModule.ts
@@ -1,24 +1,21 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { DataBaseConfigService } from '@/config/DataBaseConfigService';
 import { ConfigModule } from '@nestjs/config';
 import database from '@/config/database';
-import { CoreModule } from '@/core/CoreModule';
-import { TodoModule } from '@/todo/TodoModule';
-import { TypeOrmConfigService } from './TypeOrmConfigService';
 
 @Module({
   imports: [
-    CoreModule,
-    TodoModule,
     ConfigModule.forRoot({
+      // must be global for any module to use load typeorm repository correctly
       isGlobal: true,
       load: [database],
     }),
     TypeOrmModule.forRootAsync({
-      useClass: TypeOrmConfigService,
+      useClass: DataBaseConfigService,
     }),
   ],
   controllers: [],
   providers: [],
 })
-export class AppModule {}
+export class AppConfigModule {}
